fixes uint to ushort truncation issue for Amount
Currently the server tracks item amounts via an int or uint. The network packets, however, expect this value to be transmitted in a 2 byte ushort. The casting truncates the 16 high order bits, causing the client to receive a bizare number. This fix is simple in that it will just pass the max ushort value if the server side Amount field is too high. This really only is an issue when dealing with currencies such as Gold or Tokens.