Point2D implements ISpanFormattable
Point2D implements a basic TryFormat function. Only a single format is
supported: "(X, Y)" where X and Y are base 10 integers.

Uses recent improvements in string interpolation to write the characters
to the destination without first allocating memory for boxed arguments
or intermediate strings.

This is a partial solution to issue #1067. A similar implementation can
be implemented for other classes in Geometry if it looks promising.